+++ index.html
@@ -226,7 +226,7 @@
       </div>
     </section>
   </template>
-
+  <script src="js/main.js" type="module"></script>
 </body>
-<script src="js/main.js"></script>
+
 </html>
